refactor(tests): split test_headless.py into 4 focused modules (P1-F01 + 4 fixes)
Split tests/execution/test_headless.py (4,520 lines, 49 classes) into:
- test_headless_core.py (29 classes)
- test_headless_synthesis.py (7 classes)
- test_headless_path_validation.py (11 classes)
- test_headless_dispatch.py (2 classes + 1 standalone)

In-place fixes applied during split:
- P2-F15: Parametrize 4 identical schema-key assertions in TestBuildSkillResultCrossValidation
- P3-F18: Assert return value of datetime.fromisoformat(record.timestamp)
- P2-F08: Rename duplicate make_session fixture → make_headless_session
- P3-F07: Extract duplicate _make_success_stdout to module-level in dispatch file

Adds tests/arch/test_headless_split.py as structural guard.
