+++ tests/execution/test_headless_core.py
@@ -926,48 +926,38 @@ class TestBuildSkillResultCrossValidation:
         response = json.loads(_build_skill_result(result_obj).to_json())
         assert response["success"] is False
 
-    def test_stale_schema(self):
-        """Stale response has standard keys."""
-        result_obj = SubprocessResult(
-            returncode=-1, stdout="", stderr="", termination=TerminationReason.STALE, pid=1
-        )
-        response = json.loads(_build_skill_result(result_obj).to_json())
-        assert set(response.keys()) == self.EXPECTED_SKILL_KEYS
-
-    def test_timeout_schema(self):
-        """Timeout response has standard keys."""
-        result_obj = SubprocessResult(
-            returncode=-1, stdout="", stderr="", termination=TerminationReason.TIMED_OUT, pid=1
-        )
-        response = json.loads(_build_skill_result(result_obj).to_json())
-        assert set(response.keys()) == self.EXPECTED_SKILL_KEYS
-
-    def test_normal_success_schema(self):
-        """Normal success response has standard keys."""
-        valid_json = json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "is_error": False,
-                "result": "Done.",
-                "session_id": "s1",
-            }
-        )
-        result_obj = SubprocessResult(
-            returncode=0,
-            stdout=valid_json,
-            stderr="",
-            termination=TerminationReason.NATURAL_EXIT,
-            pid=1,
-        )
-        response = json.loads(_build_skill_result(result_obj).to_json())
-        assert set(response.keys()) == self.EXPECTED_SKILL_KEYS
-
-    def test_empty_stdout_schema(self):
-        """Empty stdout response has standard keys."""
-        result_obj = SubprocessResult(
-            returncode=0, stdout="", stderr="", termination=TerminationReason.NATURAL_EXIT, pid=1
-        )
+    @pytest.mark.parametrize(
+        "result_obj",
+        [
+            SubprocessResult(
+                returncode=-1, stdout="", stderr="", termination=TerminationReason.STALE, pid=1
+            ),
+            SubprocessResult(
+                returncode=-1, stdout="", stderr="", termination=TerminationReason.TIMED_OUT, pid=1
+            ),
+            SubprocessResult(
+                returncode=0,
+                stdout=json.dumps(
+                    {
+                        "type": "result",
+                        "subtype": "success",
+                        "is_error": False,
+                        "result": "Done.",
+                        "session_id": "s1",
+                    }
+                ),
+                stderr="",
+                termination=TerminationReason.NATURAL_EXIT,
+                pid=1,
+            ),
+            SubprocessResult(
+                returncode=0, stdout="", stderr="", termination=TerminationReason.NATURAL_EXIT, pid=1
+            ),
+        ],
+        ids=["stale", "timeout", "normal_success", "empty_stdout"],
+    )
+    def test_schema_keys(self, result_obj: SubprocessResult):
+        """Response always exposes the full standard key set."""
         response = json.loads(_build_skill_result(result_obj).to_json())
         assert set(response.keys()) == self.EXPECTED_SKILL_KEYS
 
@@ -1997,7 +1987,7 @@ class TestFailureCaptureInBuildSkillResult:
         _build_skill_result(result, skill_command="/test", audit=tool_ctx.audit)
         record = tool_ctx.audit.get_report()[0]
         assert record.timestamp  # non-empty ISO timestamp
-        datetime.fromisoformat(record.timestamp)  # must parse as ISO
+        assert datetime.fromisoformat(record.timestamp)  # valid ISO 8601 format
 
     def test_stale_termination_is_captured(self, tool_ctx):
         result = _make_result(returncode=0, termination_reason=TerminationReason.STALE)
@@ -2496,2000 +2486,3 @@ class TestInjectNarrationSuppression:
         result = _inject_narration_suppression("cmd")
         # Must not suppress the final response where structured output tokens live
         assert "final response" in result
-
-
-# ---------------------------------------------------------------------------
-# Test: _extract_output_paths (Step 1b)
-# ---------------------------------------------------------------------------
-
-
-class TestExtractOutputPaths:
-    def test_extracts_single_path(self):
-        from autoskillit.execution.headless import _extract_output_paths
-
-        msgs = ["plan_path = /correct/path/.autoskillit/temp/make-plan/foo.md"]
-        result = _extract_output_paths(msgs)
-        assert result == {"plan_path": "/correct/path/.autoskillit/temp/make-plan/foo.md"}
-
-    def test_extracts_multiple_tokens(self):
-        from autoskillit.execution.headless import _extract_output_paths
-
-        msg = (
-            "plan_path = /clone/.autoskillit/temp/make-plan/plan.md\n"
-            "summary_path = /clone/.autoskillit/temp/report/summary.md\n"
-            "investigation_path = /clone/.autoskillit/temp/investigate/inv.md"
-        )
-        result = _extract_output_paths([msg])
-        assert result == {
-            "plan_path": "/clone/.autoskillit/temp/make-plan/plan.md",
-            "summary_path": "/clone/.autoskillit/temp/report/summary.md",
-            "investigation_path": "/clone/.autoskillit/temp/investigate/inv.md",
-        }
-
-    def test_returns_empty_when_no_tokens(self):
-        from autoskillit.execution.headless import _extract_output_paths
-
-        result = _extract_output_paths(["no tokens here", "just regular text"])
-        assert result == {}
-
-    def test_ignores_non_absolute_paths(self):
-        from autoskillit.execution.headless import _extract_output_paths
-
-        result = _extract_output_paths(["plan_path = .autoskillit/temp/make-plan/foo.md"])
-        assert result == {}
-
-    def test_extracts_from_multiple_messages(self):
-        from autoskillit.execution.headless import _extract_output_paths
-
-        msgs = [
-            "plan_path = /first/path",
-            "investigation_path = /second/path",
-        ]
-        result = _extract_output_paths(msgs)
-        assert result == {
-            "plan_path": "/first/path",
-            "investigation_path": "/second/path",
-        }
-
-    def test_last_occurrence_wins(self):
-        from autoskillit.execution.headless import _extract_output_paths
-
-        msgs = [
-            "plan_path = /first/path",
-            "plan_path = /second/path",
-        ]
-        result = _extract_output_paths(msgs)
-        assert result == {"plan_path": "/second/path"}
-
-
-# ---------------------------------------------------------------------------
-# Test: _validate_output_paths (Step 1c)
-# ---------------------------------------------------------------------------
-
-
-class TestValidateOutputPaths:
-    def test_returns_none_when_all_paths_under_cwd(self):
-        from autoskillit.execution.headless import _validate_output_paths
-
-        paths = {
-            "plan_path": "/clone/path/.autoskillit/temp/make-plan/foo.md",
-            "report_path": "/clone/path/.autoskillit/temp/report/bar.md",
-        }
-        result = _validate_output_paths(paths, "/clone/path")
-        assert result is None
-
-    def test_returns_diagnostic_when_path_outside_cwd(self):
-        from autoskillit.execution.headless import _validate_output_paths
-
-        paths = {
-            "plan_path": "/source/repo/.autoskillit/temp/make-plan/foo.md",
-        }
-        result = _validate_output_paths(paths, "/clone/path")
-        assert result is not None
-        assert "plan_path" in result
-        assert "/source/repo/.autoskillit/temp/make-plan/foo.md" in result
-        assert "/clone/path" in result
-
-    def test_returns_none_for_empty_paths(self):
-        from autoskillit.execution.headless import _validate_output_paths
-
-        result = _validate_output_paths({}, "/clone/path")
-        assert result is None
-
-    def test_cwd_trailing_slash_handling(self):
-        from autoskillit.execution.headless import _validate_output_paths
-
-        paths = {"plan_path": "/clone/path/.autoskillit/temp/foo.md"}
-        assert _validate_output_paths(paths, "/clone/path/") is None
-        assert _validate_output_paths(paths, "/clone/path") is None
-
-    def test_multiple_violations(self):
-        from autoskillit.execution.headless import _validate_output_paths
-
-        paths = {
-            "plan_path": "/wrong/.autoskillit/temp/plan.md",
-            "report_path": "/wrong/.autoskillit/temp/report.md",
-        }
-        result = _validate_output_paths(paths, "/clone")
-        assert result is not None
-        assert "plan_path" in result
-        assert "report_path" in result
-
-
-# ---------------------------------------------------------------------------
-# Test: _build_skill_result path contamination detection (Step 1d)
-# ---------------------------------------------------------------------------
-
-
-class TestBuildSkillResultPathContamination:
-    @staticmethod
-    def _assistant_ndjson(text: str) -> str:
-        """Build an NDJSON assistant message line."""
-        return json.dumps(
-            {
-                "type": "assistant",
-                "message": {"content": [{"text": text}]},
-            }
-        )
-
-    def test_path_contamination_detected(self):
-        """Output paths outside cwd override success to False."""
-        path = "/wrong/source/repo/.autoskillit/temp/make-plan/foo.md"
-        stdout = (
-            self._assistant_ndjson(f"plan_path = {path}")
-            + "\n"
-            + _success_session_json("Plan created.")
-        )
-        result = _sr(0, stdout, "", TerminationReason.NATURAL_EXIT)
-        sr = _build_skill_result(result, cwd="/correct/clone")
-        assert sr.success is False
-        assert sr.subtype == "path_contamination"
-        assert sr.needs_retry is True
-        assert sr.retry_reason == RetryReason.PATH_CONTAMINATION
-
-    def test_path_contamination_emits_path_contamination_reason(self):
-        """Path contamination must emit PATH_CONTAMINATION, not RESUME.
-
-        PATH_CONTAMINATION is not a context limit — it is a CWD boundary violation.
-        The orchestrator must route to on_failure, not on_context_limit.
-        """
-        path = "/wrong/source/repo/.autoskillit/temp/make-plan/bar.md"
-        stdout = (
-            self._assistant_ndjson(f"plan_path = {path}")
-            + "\n"
-            + _success_session_json("Plan created.")
-        )
-        result = _sr(0, stdout, "", TerminationReason.NATURAL_EXIT)
-        sr = _build_skill_result(result, cwd="/correct/clone")
-        assert sr.retry_reason == RetryReason.PATH_CONTAMINATION
-        assert sr.needs_retry is True
-        # Confirm it is NOT RESUME — routing must not go to on_context_limit
-
-    def test_no_contamination_when_paths_under_cwd(self):
-        """All output paths under cwd yields normal result."""
-        path = "/correct/clone/.autoskillit/temp/make-plan/foo.md"
-        stdout = (
-            self._assistant_ndjson(f"plan_path = {path}")
-            + "\n"
-            + _success_session_json("Plan created.")
-        )
-        result = _sr(0, stdout, "", TerminationReason.NATURAL_EXIT)
-        sr = _build_skill_result(result, cwd="/correct/clone")
-        assert sr.success is True
-        assert sr.subtype != "path_contamination"
-
-    def test_no_contamination_when_cwd_empty(self):
-        """Empty cwd skips path validation — direct callers without clone context."""
-        path = "/any/path/.autoskillit/temp/foo.md"
-        stdout = (
-            self._assistant_ndjson(f"plan_path = {path}") + "\n" + _success_session_json("Done.")
-        )
-        result = _sr(0, stdout, "", TerminationReason.NATURAL_EXIT)
-        sr = _build_skill_result(result, cwd="")
-        assert sr.success is True
-
-    def test_no_contamination_when_no_path_tokens(self):
-        """No output path tokens means validation passes."""
-        stdout = _success_session_json("Done with no file output.")
-        result = _sr(0, stdout, "", TerminationReason.NATURAL_EXIT)
-        sr = _build_skill_result(result, cwd="/clone/path")
-        assert sr.success is True
-
-
-# ---------------------------------------------------------------------------
-# Test: run_headless_core passes cwd (Step 1f)
-# ---------------------------------------------------------------------------
-
-
-class TestRunHeadlessCorePassesCwd:
-    @pytest.mark.anyio
-    async def test_cwd_anchor_injected_into_prompt(self, tool_ctx):
-        """Verify run_headless_core injects cwd anchor into the skill prompt."""
-        from autoskillit.execution.headless import run_headless_core
-
-        payload = _success_session_json("Result text %%ORDER_UP%%")
-        tool_ctx.runner.push(
-            SubprocessResult(0, payload, "", TerminationReason.NATURAL_EXIT, pid=1)
-        )
-        await run_headless_core(
-            "/autoskillit:investigate test",
-            "/some/test/cwd",
-            tool_ctx,
-        )
-        assert tool_ctx.runner.call_args_list, "Runner was never called"
-        last_cmd = tool_ctx.runner.call_args_list[-1][0]
-        # cmd is ["env", ...vars, "claude", "-p", <prompt>, ...]
-        p_idx = last_cmd.index("-p")
-        prompt_arg = last_cmd[p_idx + 1]
-        assert "WORKING DIRECTORY ANCHOR" in prompt_arg
-        assert "/some/test/cwd" in prompt_arg
-
-
-def _make_tool_use_line(name: str, input_dict: dict) -> str:
-    return json.dumps(
-        {
-            "type": "assistant",
-            "message": {
-                "content": [{"type": "tool_use", "name": name, "id": "x", "input": input_dict}]
-            },
-        }
-    )
-
-
-class TestScanJsonlWritePaths:
-    CWD = "/clone/worktree"
-
-    def test_returns_empty_for_clean_write_inside_cwd(self):
-        line = _make_tool_use_line(
-            "Write", {"file_path": f"{self.CWD}/.autoskillit/temp/out.md", "content": "x"}
-        )
-        assert _scan_jsonl_write_paths(line, self.CWD) == []
-
-    def test_detects_write_outside_cwd(self):
-        line = _make_tool_use_line(
-            "Write", {"file_path": "/source/repo/.autoskillit/temp/stolen.md", "content": "x"}
-        )
-        warnings = _scan_jsonl_write_paths(line, self.CWD)
-        assert len(warnings) == 1
-        assert "/source/repo/.autoskillit/temp/stolen.md" in warnings[0]
-
-    def test_detects_edit_outside_cwd(self):
-        line = _make_tool_use_line(
-            "Edit",
-            {
-                "file_path": "/source/repo/src/autoskillit/file.py",
-                "old_string": "a",
-                "new_string": "b",
-            },
-        )
-        warnings = _scan_jsonl_write_paths(line, self.CWD)
-        assert len(warnings) == 1
-        assert "Edit" in warnings[0]
-
-    def test_detects_bash_with_absolute_path_outside_cwd(self):
-        line = _make_tool_use_line(
-            "Bash", {"command": "cat /source/repo/README.md > /tmp/out.txt"}
-        )
-        warnings = _scan_jsonl_write_paths(line, self.CWD)
-        assert len(warnings) >= 1
-        assert any("/source/repo" in w for w in warnings)
-
-    def test_no_warnings_for_empty_stdout(self):
-        assert _scan_jsonl_write_paths("", self.CWD) == []
-
-    def test_no_warnings_for_malformed_jsonl(self):
-        assert _scan_jsonl_write_paths("not json at all\n{broken", self.CWD) == []
-
-    def test_no_warnings_for_read_only_tool_calls(self):
-        line = _make_tool_use_line("Read", {"file_path": "/source/repo/some_file.py"})
-        assert _scan_jsonl_write_paths(line, self.CWD) == []
-
-    def test_multiple_violations_in_one_session(self):
-        lines = "\n".join(
-            [
-                _make_tool_use_line("Write", {"file_path": "/source/repo/a.md", "content": "x"}),
-                _make_tool_use_line(
-                    "Edit",
-                    {"file_path": "/source/repo/b.py", "old_string": "a", "new_string": "b"},
-                ),
-                json.dumps(
-                    {
-                        "type": "result",
-                        "subtype": "success",
-                        "result": "done",
-                        "session_id": "",
-                        "is_error": False,
-                    }
-                ),
-            ]
-        )
-        warnings = _scan_jsonl_write_paths(lines, self.CWD)
-        assert len(warnings) == 2
-
-    def test_no_warning_when_cwd_empty(self):
-        line = _make_tool_use_line("Write", {"file_path": "/any/path/file.md", "content": "x"})
-        assert _scan_jsonl_write_paths(line, "") == []
-
-    def test_no_warning_when_cwd_is_relative(self):
-        line = _make_tool_use_line("Write", {"file_path": "/any/path/file.md", "content": "x"})
-        assert _scan_jsonl_write_paths(line, "relative/path") == []
-
-
-class TestBuildSkillResultWritePathWarnings:
-    def test_write_path_warnings_empty_for_clean_session(self):
-        stdout = (
-            _make_tool_use_line(
-                "Write", {"file_path": "/clone/worktree/.autoskillit/temp/out.md", "content": "x"}
-            )
-            + "\n"
-            + _success_session_json("Done %%DONE%%")
-        )
-        result = _sr(0, stdout, "", TerminationReason.NATURAL_EXIT)
-        sr = _build_skill_result(result, cwd="/clone/worktree")
-        assert sr.write_path_warnings == []
-
-    def test_write_path_warnings_populated_for_contaminated_session(self):
-        stdout = (
-            _make_tool_use_line(
-                "Write", {"file_path": "/source/repo/.autoskillit/temp/stolen.md", "content": "x"}
-            )
-            + "\n"
-            + _success_session_json("Done %%DONE%%")
-        )
-        result = _sr(0, stdout, "", TerminationReason.NATURAL_EXIT)
-        sr = _build_skill_result(result, cwd="/clone/worktree")
-        assert len(sr.write_path_warnings) == 1
-        assert "/source/repo/.autoskillit/temp/stolen.md" in sr.write_path_warnings[0]
-
-    def test_write_path_warnings_appear_in_to_json(self):
-        stdout = (
-            _make_tool_use_line("Write", {"file_path": "/source/repo/bad.md", "content": "x"})
-            + "\n"
-            + _success_session_json("Done %%DONE%%")
-        )
-        result = _sr(0, stdout, "", TerminationReason.NATURAL_EXIT)
-        sr = _build_skill_result(result, cwd="/clone/worktree")
-        data = json.loads(sr.to_json())
-        assert "write_path_warnings" in data
-        assert len(data["write_path_warnings"]) == 1
-
-    def test_write_path_warnings_independent_of_output_token_contamination(self):
-        """Warnings are populated even when _validate_output_paths also fires."""
-        # plan_path token must appear in an assistant text message for
-        # _validate_output_paths to detect it (it scans assistant_messages,
-        # not the final result record).
-        path_token_line = json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "content": [
-                        {
-                            "type": "text",
-                            "text": "plan_path = /source/repo/.autoskillit/temp/plan.md",
-                        }
-                    ]
-                },
-            }
-        )
-        stdout = (
-            _make_tool_use_line("Write", {"file_path": "/source/repo/bad.md", "content": "x"})
-            + "\n"
-            + path_token_line
-            + "\n"
-            + _success_session_json("Done %%DONE%%")
-        )
-        result = _sr(0, stdout, "", TerminationReason.NATURAL_EXIT)
-        sr = _build_skill_result(result, cwd="/clone/worktree")
-        # subtype is path_contamination (from _validate_output_paths)
-        assert sr.subtype == "path_contamination"
-        # write_path_warnings also populated from JSONL scan
-        assert len(sr.write_path_warnings) >= 1
-
-
-class TestBuildSkillResultChannelBPatternRecovery:
-    """When Channel B wins and pattern is absent from result but present in
-    assistant_messages, _build_skill_result should recover and produce success=True.
-    """
-
-    def test_build_skill_result_channel_b_recovers_pattern_from_assistant_messages(
-        self,
-    ) -> None:
-        """Channel B wins before stdout drain; pattern found in assistant_messages
-        → recovery produces success=True with block in result.
-        """
-        block = "---prepare-issue-result---\n{}\n---/prepare-issue-result---"
-        # Assistant message has the block but NOT a standalone %%ORDER_UP%% marker,
-        # so _recover_from_separate_marker will not activate. The new
-        # _recover_block_from_assistant_messages must find the block here instead.
-        assistant_line = json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "content": [{"type": "text", "text": block}],
-                },
-            }
-        )
-        result_line = json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "is_error": False,
-                "result": "",  # stdout not yet drained — Channel B won the race
-                "session_id": "s1",
-                "errors": [],
-            }
-        )
-        stdout = assistant_line + "\n" + result_line
-
-        sub_result = SubprocessResult(
-            returncode=0,
-            stdout=stdout,
-            stderr="",
-            termination=TerminationReason.COMPLETED,
-            pid=12345,
-            channel_confirmation=ChannelConfirmation.CHANNEL_B,
-        )
-        sr = _build_skill_result(
-            sub_result,
-            completion_marker="%%ORDER_UP%%",
-            expected_output_patterns=["---prepare-issue-result---"],
-        )
-        assert sr.success is True
-        assert "---prepare-issue-result---" in sr.result
-
-    def test_synthesis_not_run_for_unparseable_channel_b_with_write_evidence(self) -> None:
-        """CHANNEL_B + UNPARSEABLE + write evidence must not produce success=True.
-
-        Recovery (and CHANNEL_B bypass) must be blocked when session.session_complete
-        is False. Synthesis must not inject the file path into result.
-        """
-
-        tool_use_line = json.dumps(
-            {
-                "type": "tool_use",
-                "id": "t1",
-                "name": "Write",
-                "input": {
-                    "file_path": "/cwd/.autoskillit/temp/make-plan/arch_lens_selection_2026-01-01.md"  # noqa: E501
-                },
-            }
-        )
-        # UNPARSEABLE result (no proper result record)
-        raw_stdout = tool_use_line + "\ngarbage ndjson partial"
-
-        sub_result = SubprocessResult(
-            returncode=1,
-            stdout=raw_stdout,
-            stderr="",
-            termination=TerminationReason.COMPLETED,
-            pid=12345,
-            channel_confirmation=ChannelConfirmation.CHANNEL_B,
-        )
-        sr = _build_skill_result(
-            sub_result,
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-        )
-        assert sr.success is False
-        assert sr.subtype not in {"success"}
-        # Synthesis must not have injected the structured token (raw NDJSON may contain
-        # the file path string, but the plan_path = ... assignment must be absent)
-        assert "plan_path =" not in sr.result
-
-    def test_synthesis_not_run_for_timeout_channel_b_with_write_evidence(self) -> None:
-        """CHANNEL_B + TIMEOUT + write evidence must not produce success=True."""
-        tool_use_line = json.dumps(
-            {
-                "type": "tool_use",
-                "id": "t1",
-                "name": "Write",
-                "input": {
-                    "file_path": "/cwd/.autoskillit/temp/make-plan/arch_lens_selection_2026-01-01.md"  # noqa: E501
-                },
-            }
-        )
-        result_line = json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "is_error": False,
-                "result": "",
-                "session_id": "s1",
-                "errors": [],
-            }
-        )
-        stdout = tool_use_line + "\n" + result_line
-
-        sub_result = SubprocessResult(
-            returncode=-1,
-            stdout=stdout,
-            stderr="",
-            termination=TerminationReason.TIMED_OUT,
-            pid=12345,
-            channel_confirmation=ChannelConfirmation.CHANNEL_B,
-        )
-        sr = _build_skill_result(
-            sub_result,
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-        )
-        assert sr.success is False
-        assert "arch_lens_selection" not in sr.result
-
-    def test_synthesis_skipped_for_channel_b_session_complete(self) -> None:
-        """CHANNEL_B + SUCCESS + write evidence but no pattern in assistant_messages.
-
-        Synthesis must NOT fabricate the token — if pattern is absent from
-        assistant_messages, the agent never emitted it.
-        """
-        tool_use_line = json.dumps(
-            {
-                "type": "tool_use",
-                "id": "t1",
-                "name": "Write",
-                "input": {"file_path": "/cwd/.autoskillit/temp/make-plan/task_plan.md"},
-            }
-        )
-        result_line = json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "is_error": False,
-                "result": "done",
-                "session_id": "s1",
-                "errors": [],
-            }
-        )
-        stdout = tool_use_line + "\n" + result_line
-
-        sub_result = SubprocessResult(
-            returncode=0,
-            stdout=stdout,
-            stderr="",
-            termination=TerminationReason.COMPLETED,
-            pid=12345,
-            channel_confirmation=ChannelConfirmation.CHANNEL_B,
-        )
-        sr = _build_skill_result(
-            sub_result,
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-        )
-        assert sr.success is False
-        assert "plan_path" not in sr.result
-
-
-class TestBuildSkillResultChannelAPatternRecovery:
-    """_build_skill_result must extend pattern recovery to CHANNEL_A wins, not just CHANNEL_B."""
-
-    def test_build_skill_result_channel_a_recovers_from_assistant_messages(self) -> None:
-        """_build_skill_result must attempt pattern recovery for CHANNEL_A, not just CHANNEL_B."""
-        block = "verdict = GO\n%%ORDER_UP%%"
-        assistant_line = json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "content": [{"type": "text", "text": block}],
-                },
-            }
-        )
-        result_line = json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "is_error": False,
-                "result": "Done. %%ORDER_UP%%",
-                "session_id": "s1",
-                "errors": [],
-            }
-        )
-        stdout = assistant_line + "\n" + result_line
-        sub_result = SubprocessResult(
-            returncode=0,
-            stdout=stdout,
-            stderr="",
-            termination=TerminationReason.NATURAL_EXIT,
-            pid=12345,
-            channel_confirmation=ChannelConfirmation.CHANNEL_A,
-        )
-        sr = _build_skill_result(
-            sub_result,
-            completion_marker="%%ORDER_UP%%",
-            expected_output_patterns=["verdict\\s*=\\s*(GO|NO GO)"],
-        )
-        assert sr.success is True, (
-            "CHANNEL_A should recover patterns from assistant_messages, same as CHANNEL_B."
-        )
-
-    def test_build_skill_result_channel_a_pattern_contract_violation_is_terminal(
-        self,
-    ) -> None:
-        """After failed CHANNEL_A recovery, contract violation must be FAILED, not RETRIABLE."""
-        assistant_line = json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "content": [{"type": "text", "text": "No verdict here. %%ORDER_UP%%"}],
-                },
-            }
-        )
-        result_line = json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "is_error": False,
-                "result": "Done. %%ORDER_UP%%",
-                "session_id": "s1",
-                "errors": [],
-            }
-        )
-        stdout = assistant_line + "\n" + result_line
-        sub_result = SubprocessResult(
-            returncode=0,
-            stdout=stdout,
-            stderr="",
-            termination=TerminationReason.NATURAL_EXIT,
-            pid=12345,
-            channel_confirmation=ChannelConfirmation.CHANNEL_A,
-        )
-        sr = _build_skill_result(
-            sub_result,
-            completion_marker="%%ORDER_UP%%",
-            expected_output_patterns=["verdict\\s*=\\s*(GO|NO GO)"],
-        )
-        assert sr.success is False
-        assert sr.needs_retry is False
-        assert sr.subtype == "adjudicated_failure"
-
-    def test_bold_wrapped_plan_path_returns_success_not_adjudicated_failure(
-        self,
-    ) -> None:
-        """Regression test for issue #462: model emitting **plan_path** = /abs/path
-        must produce success=True, not adjudicated_failure."""
-        result_text = (
-            "The implementation plan has been written.\n\n"
-            "**plan_path** = /abs/path/plan.md\n"
-            "**plan_parts** = /abs/path/plan.md\n"
-            "%%ORDER_UP%%"
-        )
-        sub_result = SubprocessResult(
-            returncode=0,
-            stdout=_success_session_json(result_text),
-            stderr="",
-            termination=TerminationReason.NATURAL_EXIT,
-            pid=12345,
-        )
-        sr = _build_skill_result(
-            sub_result,
-            completion_marker="%%ORDER_UP%%",
-            expected_output_patterns=["plan_path\\s*=\\s*/.+"],
-        )
-        assert sr.success is True
-        assert sr.subtype != "adjudicated_failure"
-        assert sr.needs_retry is False
-
-    def test_bold_wrapped_tokens_in_assistant_messages_recovery_succeeds(
-        self,
-    ) -> None:
-        """Recovery path: bold-wrapped tokens in assistant_messages must also be found.
-        The pattern_recovery block in _build_skill_result requires channel_confirmation
-        != UNMONITORED to trigger; use CHANNEL_A to activate the recovery path."""
-        token_block = "**plan_path** = /abs/path/plan.md\n**plan_parts** = /abs/path/plan.md"
-        assistant_line = json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "content": [{"type": "text", "text": token_block}],
-                },
-            }
-        )
-        result_line = json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "is_error": False,
-                "result": "The plan is complete. %%ORDER_UP%%",
-                "session_id": "s1",
-                "errors": [],
-            }
-        )
-        stdout = assistant_line + "\n" + result_line
-        sub_result = SubprocessResult(
-            returncode=0,
-            stdout=stdout,
-            stderr="",
-            termination=TerminationReason.NATURAL_EXIT,
-            pid=12345,
-            channel_confirmation=ChannelConfirmation.CHANNEL_A,
-        )
-        sr = _build_skill_result(
-            sub_result,
-            completion_marker="%%ORDER_UP%%",
-            expected_output_patterns=["plan_path\\s*=\\s*/.+"],
-        )
-        assert sr.success is True
-        assert sr.subtype != "adjudicated_failure"
-        assert sr.needs_retry is False
-
-
-class TestBuildSkillResultDirMissingRecovery:
-    """DIR_MISSING channel confirmation does not silently pass — it attempts
-    late-bind recovery when conditions allow."""
-
-    def test_dir_missing_with_recoverable_subtype_attempts_recovery(self):
-        """When channel_confirmation is DIR_MISSING and subtype is recoverable,
-        the recovery gate must attempt marker-based recovery.
-
-        Setup: termination=COMPLETED (bypasses stale branch), subtype=empty_output
-        (in _CHANNEL_B_RECOVERABLE_SUBTYPES), marker on a standalone line in the
-        assistant message (required by _marker_is_standalone).
-        """
-        import structlog.testing
-
-        marker = "===DONE==="
-        # Marker must appear as a standalone line for _marker_is_standalone to match
-        assistant_line = json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "content": [{"type": "text", "text": f"Task complete.\n\n{marker}"}],
-                },
-            }
-        )
-        # subtype=empty_output places the session in _CHANNEL_B_RECOVERABLE_SUBTYPES,
-        # triggering the DIR_MISSING recovery guard in _build_skill_result
-        result_line = json.dumps(
-            {
-                "type": "result",
-                "subtype": "empty_output",
-                "is_error": True,
-                "result": "",
-                "session_id": "s1",
-                "errors": [],
-            }
-        )
-        stdout = assistant_line + "\n" + result_line
-        # termination=COMPLETED skips the stale-branch early return so execution
-        # reaches the Channel B / DIR_MISSING recovery gate
-        sub_result = SubprocessResult(
-            returncode=0,
-            stdout=stdout,
-            stderr="",
-            termination=TerminationReason.COMPLETED,
-            pid=12345,
-            channel_confirmation=ChannelConfirmation.DIR_MISSING,
-        )
-        with structlog.testing.capture_logs() as logs:
-            skill = _build_skill_result(sub_result, completion_marker=marker)
-        # Recovery should succeed — marker found as standalone line in assistant_messages
-        assert skill.success is True
-        # Verify the recovery code path was taken, not just the outcome
-        assert any(e.get("event") == "dir_missing_late_bind_recovery" for e in logs)
-
-    def test_dir_missing_without_marker_does_not_recover(self):
-        """DIR_MISSING with no completion marker must not silently succeed."""
-        sub_result = SubprocessResult(
-            returncode=0,
-            stdout="",
-            stderr="",
-            termination=TerminationReason.STALE,
-            pid=12345,
-            channel_confirmation=ChannelConfirmation.DIR_MISSING,
-        )
-        skill = _build_skill_result(sub_result)
-        assert skill.success is False
-
-
-class TestTimedOutSessionPreservesState:
-    """TIMED_OUT branch must parse stdout to preserve tool_uses and assistant_messages."""
-
-    def test_timed_out_with_writes_preserves_write_call_count(self):
-        """Timed-out session with Write/Edit tool_use blocks must report write_call_count > 0."""
-        from autoskillit.execution.headless import _build_skill_result
-
-        ndjson = "\n".join(
-            [
-                _make_tool_use_line("Write", {"file_path": "/tmp/a.py", "content": "x"}),
-                _make_tool_use_line(
-                    "Edit", {"file_path": "/tmp/b.py", "old_string": "a", "new_string": "b"}
-                ),
-            ]
-        )
-        sub_result = SubprocessResult(
-            returncode=-1,
-            stdout=ndjson,
-            stderr="",
-            termination=TerminationReason.TIMED_OUT,
-            pid=12345,
-        )
-        sr = _build_skill_result(sub_result)
-        assert sr.write_call_count == 2
-        assert sr.success is False
-        assert sr.needs_retry is False
-
-    def test_timed_out_with_empty_stdout_uses_timeout_subtype(self):
-        """Timed-out session with no stdout uses TIMEOUT subtype."""
-        from autoskillit.execution.headless import _build_skill_result
-
-        sub_result = SubprocessResult(
-            returncode=-1,
-            stdout="",
-            stderr="",
-            termination=TerminationReason.TIMED_OUT,
-            pid=12345,
-        )
-        sr = _build_skill_result(sub_result)
-        assert sr.success is False
-        assert sr.write_call_count == 0
-
-    def test_timed_out_with_success_result_overrides_to_timeout(self):
-        """When timed-out stdout has a success result, subtype is overridden to timeout."""
-        from autoskillit.execution.headless import _build_skill_result
-
-        ndjson = "\n".join(
-            [
-                _make_tool_use_line("Write", {"file_path": "/tmp/a.py", "content": "x"}),
-                json.dumps(
-                    {
-                        "type": "result",
-                        "subtype": "success",
-                        "is_error": False,
-                        "result": "Task completed.",
-                        "session_id": "s1",
-                    }
-                ),
-            ]
-        )
-        sub_result = SubprocessResult(
-            returncode=-1,
-            stdout=ndjson,
-            stderr="",
-            termination=TerminationReason.TIMED_OUT,
-            pid=12345,
-        )
-        sr = _build_skill_result(sub_result)
-        assert sr.cli_subtype == "timeout"
-        assert sr.write_call_count == 1
-
-
-class TestOutputPathTokensDerivedFromContracts:
-    # Hardcoded fixture — update when skill_contracts.yaml gains new file_path outputs.
-    # Using a fixed set (rather than re-deriving) ensures bugs in the derivation formula
-    # cause test failures rather than silent agreement between production and test code.
-    _EXPECTED_OUTPUT_PATH_TOKENS = frozenset(
-        {
-            "analysis_file",
-            "campaign_path",
-            "conflict_report_path",
-            "diagnosis_path",
-            "diagram_path",
-            "evaluation_dashboard",
-            "experiment_plan",
-            "group_files",
-            "groups_path",
-            "investigation_path",
-            "manifest_path",
-            "plan_parts",
-            "plan_path",
-            "pr_order_file",
-            "prep_path",
-            "recipe_path",
-            "remediation_path",
-            "report_path",
-            "report_plan_path",
-            "results_path",
-            "review_path",
-            "revision_guidance",
-            "scope_report",
-            "summary_path",
-            "triage_manifest",
-            "triage_report",
-            "visualization_plan_path",
-            "html_path",
-            "resource_report",
-            "execution_map",
-            "execution_map_report",
-            # planner-generate-phases output (planner recipe)
-            "phase_manifest_path",
-            # planner-elaborate-phase output (parallel worker)
-            "elab_result_path",
-            # planner-refine-phases output
-            "refined_plan_path",
-            # planner-refine-assignments output
-            "refined_assignments_path",
-            # planner-refine-wps output
-            "refined_wps_path",
-        }
-    )
-
-    def test_output_path_tokens_matches_expected_fixture(self) -> None:
-        """_OUTPUT_PATH_TOKENS must exactly match the known fixture set.
-
-        This guards against bugs in the derivation formula: if _build_path_token_set()
-        is broken, both the production frozenset and a re-derived set would agree, but
-        this hardcoded fixture would not.
-        """
-        from autoskillit.execution.headless import _OUTPUT_PATH_TOKENS
-
-        extra = _OUTPUT_PATH_TOKENS - self._EXPECTED_OUTPUT_PATH_TOKENS
-        missing = self._EXPECTED_OUTPUT_PATH_TOKENS - _OUTPUT_PATH_TOKENS
-        assert _OUTPUT_PATH_TOKENS == self._EXPECTED_OUTPUT_PATH_TOKENS, (
-            f"_OUTPUT_PATH_TOKENS diverged from expected fixture.\n"
-            f"Extra (in production, not in fixture): {extra}\n"
-            f"Missing (in fixture, not in production): {missing}"
-        )
-
-
-# ---------------------------------------------------------------------------
-# T2: _synthesize_from_write_artifacts
-# ---------------------------------------------------------------------------
-
-from autoskillit.execution.session import ClaudeSessionResult  # noqa: E402
-
-
-@pytest.fixture
-def make_session():
-    """Build a ClaudeSessionResult with configurable result and tool_uses (parsed form)."""
-
-    def _factory(
-        result: str = "",
-        tool_uses: list[dict] | None = None,
-        subtype: str = "success",
-        is_error: bool = False,
-    ) -> ClaudeSessionResult:
-        return ClaudeSessionResult(
-            subtype=subtype,
-            is_error=is_error,
-            result=result,
-            session_id="test-session",
-            tool_uses=tool_uses or [],
-        )
-
-    return _factory
-
-
-class TestSynthesizeFromWriteArtifacts:
-    def test_synthesizes_plan_path_from_write_tool_use(self, make_session):
-        """When Write file_path is absolute and pattern is plan_path=, token is injected."""
-        from autoskillit.execution.headless import _synthesize_from_write_artifacts
-
-        session = make_session(
-            result="plan summary\n%%ORDER_UP%%",
-            tool_uses=[
-                {"name": "Write", "id": "t1", "file_path": "/abs/temp/make-plan/my_plan.md"}
-            ],
-        )
-        patterns = [r"plan_path\s*=\s*/.+"]
-        result = _synthesize_from_write_artifacts(session, patterns, write_call_count=1)
-        assert result is not None
-        assert "plan_path = /abs/temp/make-plan/my_plan.md" in result.result
-
-    def test_returns_none_when_no_write_tool_uses(self, make_session):
-        """No synthesis when write_call_count == 0."""
-        from autoskillit.execution.headless import _synthesize_from_write_artifacts
-
-        session = make_session(result="plan summary\n%%ORDER_UP%%", tool_uses=[])
-        result = _synthesize_from_write_artifacts(
-            session, [r"plan_path\s*=\s*/.+"], write_call_count=0
-        )
-        assert result is None
-
-    def test_returns_none_when_no_absolute_file_path(self, make_session):
-        """Relative file_path in Write tool_use is not used for synthesis."""
-        from autoskillit.execution.headless import _synthesize_from_write_artifacts
-
-        session = make_session(
-            result="plan summary\n%%ORDER_UP%%",
-            tool_uses=[{"name": "Write", "id": "t1", "file_path": "temp/make-plan/plan.md"}],
-        )
-        result = _synthesize_from_write_artifacts(
-            session, [r"plan_path\s*=\s*/.+"], write_call_count=1
-        )
-        assert result is None
-
-    def test_returns_none_when_pattern_already_satisfied(self, make_session):
-        """No synthesis when the pattern is already in session.result."""
-        from autoskillit.execution.headless import _synthesize_from_write_artifacts
-
-        session = make_session(
-            result="plan_path = /abs/plan.md\n%%ORDER_UP%%",
-            tool_uses=[{"name": "Write", "id": "t1", "file_path": "/abs/plan.md"}],
-        )
-        result = _synthesize_from_write_artifacts(
-            session, [r"plan_path\s*=\s*/.+"], write_call_count=1
-        )
-        assert result is None
-
-    def test_returns_none_when_no_path_capture_patterns(self, make_session):
-        """Non-path patterns (verdict= etc) are not attempted for synthesis."""
-        from autoskillit.execution.headless import _synthesize_from_write_artifacts
-
-        session = make_session(
-            result="%%ORDER_UP%%",
-            tool_uses=[{"name": "Write", "id": "t1", "file_path": "/abs/plan.md"}],
-        )
-        result = _synthesize_from_write_artifacts(
-            session, [r"verdict\s*=\s*(GO|NO GO)"], write_call_count=1
-        )
-        assert result is None
-
-    def test_synthesis_uses_last_write_not_first(self, make_session):
-        """When multiple Write tool_uses exist, synthesis must use the LAST absolute path.
-
-        Multi-artifact skills write intermediate files first, final deliverable last.
-        Synthesis must inject the final deliverable (last path), not the intermediate.
-        """
-        from autoskillit.execution.headless import _synthesize_from_write_artifacts
-
-        session = make_session(
-            result="",
-            tool_uses=[
-                {
-                    "name": "Write",
-                    "id": "t1",
-                    "file_path": "/cwd/.autoskillit/temp/make-plan/arch_lens_selection.md",
-                },
-                {
-                    "name": "Write",
-                    "id": "t2",
-                    "file_path": "/cwd/.autoskillit/temp/make-plan/task_plan_2026-01-01.md",
-                },
-            ],
-        )
-        result = _synthesize_from_write_artifacts(
-            session, [r"plan_path\s*=\s*/.+"], write_call_count=2
-        )
-        assert result is not None
-        assert (
-            "plan_path = /cwd/.autoskillit/temp/make-plan/task_plan_2026-01-01.md" in result.result
-        )
-        assert "arch_lens_selection" not in result.result
-
-
-# ---------------------------------------------------------------------------
-# T3: _build_skill_result integration — CONTRACT_RECOVERY gate
-# ---------------------------------------------------------------------------
-
-from autoskillit.pipeline.audit import DefaultAuditLog, FailureRecord  # noqa: E402
-
-
-@pytest.fixture
-def make_build_skill_result_kwargs():
-    """Build a dict of kwargs for _build_skill_result from high-level parameters.
-
-    tool_uses entries use the parsed form: 'file_path' at the top level (not nested
-    in 'input'). The fixture converts them to raw NDJSON tool_use content blocks
-    with 'input': {'file_path': ...} so parse_session_result (after Step 3) preserves them.
-
-    consecutive_failures_over_budget=True creates an audit with 4 recorded failures
-    (> max_consecutive_retries=3 default) so the budget guard fires.
-    """
-    from datetime import UTC, datetime
-
-    def _factory(
-        result_text: str = "done",
-        completion_marker: str = "",
-        expected_output_patterns: list[str] | None = None,
-        tool_uses: list[dict] | None = None,
-        consecutive_failures_over_budget: bool = False,
-    ) -> dict:
-        records = []
-        if tool_uses:
-            content = []
-            for tu in tool_uses:
-                block: dict = {"type": "tool_use", "name": tu["name"], "id": tu.get("id", "")}
-                fp = tu.get("file_path")
-                if fp:
-                    block["input"] = {"file_path": fp}
-                content.append(block)
-            records.append(json.dumps({"type": "assistant", "message": {"content": content}}))
-        records.append(
-            json.dumps(
-                {
-                    "type": "result",
-                    "subtype": "success",
-                    "is_error": False,
-                    "result": result_text,
-                    "session_id": "test-session",
-                }
-            )
-        )
-        ndjson = "\n".join(records)
-        result = SubprocessResult(
-            returncode=0,
-            stdout=ndjson,
-            stderr="",
-            termination=TerminationReason.NATURAL_EXIT,
-            pid=12345,
-        )
-        audit = None
-        skill_command = ""
-        if consecutive_failures_over_budget:
-            skill_command = "/autoskillit:make-plan"
-            audit = DefaultAuditLog()
-            for _ in range(4):  # 4 > max_consecutive_retries=3 → budget exhausted
-                audit.record_failure(
-                    FailureRecord(  # type: ignore[arg-type]
-                        timestamp=datetime.now(UTC).isoformat(),
-                        skill_command=skill_command,
-                        exit_code=-1,
-                        subtype="stale",
-                        needs_retry=True,
-                        retry_reason="stale",
-                        stderr="",
-                    )
-                )
-        return {
-            "result": result,
-            "completion_marker": completion_marker,
-            "expected_output_patterns": expected_output_patterns or [],
-            "skill_command": skill_command,
-            "audit": audit,
-        }
-
-    return _factory
-
-
-class TestContractRecoveryGate:
-    def test_adjudicated_failure_with_write_evidence_becomes_retriable(
-        self, make_build_skill_result_kwargs
-    ):
-        """
-        COMPLETED + marker present + plan_path absent + Write call with no file_path
-        → synthesis finds no usable path → CONTRACT_RECOVERY gate fires
-        → needs_retry=True, retry_reason=contract_recovery.
-        """
-        kwargs = make_build_skill_result_kwargs(
-            result_text="plan summary\n%%ORDER_UP%%",
-            completion_marker="%%ORDER_UP%%",
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-            tool_uses=[
-                # Write call exists (write evidence present) but no file_path key
-                # → synthesis has nothing to inject → CONTRACT_RECOVERY gate fires
-                {"name": "Write", "id": "t1"}
-            ],
-        )
-        sr = _build_skill_result(**kwargs)
-        assert sr.needs_retry is True
-        assert sr.retry_reason == RetryReason.CONTRACT_RECOVERY
-        assert sr.success is False
-
-    def test_adjudicated_failure_without_write_evidence_stays_terminal(
-        self, make_build_skill_result_kwargs
-    ):
-        """
-        COMPLETED + marker present + plan_path absent + write_call_count == 0
-        → success=False, needs_retry=False, subtype=adjudicated_failure (unchanged behavior).
-        """
-        kwargs = make_build_skill_result_kwargs(
-            result_text="plan summary\n%%ORDER_UP%%",
-            completion_marker="%%ORDER_UP%%",
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-            tool_uses=[],
-        )
-        sr = _build_skill_result(**kwargs)
-        assert sr.success is False
-        assert sr.needs_retry is False
-        assert sr.subtype == "adjudicated_failure"
-
-    def test_synthesis_succeeds_plan_path_token_from_write_file_path(
-        self, make_build_skill_result_kwargs
-    ):
-        """
-        COMPLETED + marker present + plan_path absent + Write file_path is absolute
-        → synthesis injects token → success=True (no CONTRACT_RECOVERY needed).
-        """
-        kwargs = make_build_skill_result_kwargs(
-            result_text="plan summary\n%%ORDER_UP%%",
-            completion_marker="%%ORDER_UP%%",
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-            tool_uses=[
-                {"name": "Write", "id": "t1", "file_path": "/abs/temp/make-plan/my_plan.md"}
-            ],
-        )
-        sr = _build_skill_result(**kwargs)
-        assert sr.success is True
-        assert "plan_path = /abs/temp/make-plan/my_plan.md" in sr.result
-
-    def test_contract_recovery_respects_budget_guard(self, make_build_skill_result_kwargs):
-        """
-        When budget is exhausted, budget guard must override CONTRACT_RECOVERY.
-        Write call with no file_path → synthesis fails → CONTRACT_RECOVERY promotes
-        to retriable → budget guard fires (second call inside gate) → needs_retry=False.
-        """
-        kwargs = make_build_skill_result_kwargs(
-            result_text="plan summary\n%%ORDER_UP%%",
-            completion_marker="%%ORDER_UP%%",
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-            tool_uses=[{"name": "Write", "id": "t1"}],  # no file_path → synthesis fails
-            consecutive_failures_over_budget=True,
-        )
-        sr = _build_skill_result(**kwargs)
-        assert sr.needs_retry is False
-        assert sr.retry_reason == RetryReason.BUDGET_EXHAUSTED
-
-
-class TestBuildSkillResultSessionIdFromSubprocess:
-    """_build_skill_result propagates result.session_id on all paths."""
-
-    def test_stale_path_session_id_from_subprocess_result(self) -> None:
-        """Stale path: SkillResult.session_id == result.session_id (not hardcoded '')."""
-        result = SubprocessResult(
-            returncode=-1,
-            stdout="",
-            stderr="",
-            termination=TerminationReason.STALE,
-            pid=1,
-            session_id="real-uuid-from-channel-b",
-        )
-        sr = _build_skill_result(result)
-        assert sr.session_id == "real-uuid-from-channel-b"
-
-    def test_timeout_empty_stdout_session_id_from_subprocess_result(self) -> None:
-        """TIMED_OUT with empty stdout: SkillResult.session_id == result.session_id."""
-        result = SubprocessResult(
-            returncode=-1,
-            stdout="",
-            stderr="",
-            termination=TerminationReason.TIMED_OUT,
-            pid=1,
-            session_id="real-uuid-from-channel-b",
-        )
-        sr = _build_skill_result(result)
-        assert sr.session_id == "real-uuid-from-channel-b"
-
-    def test_channel_a_session_id_takes_precedence(self) -> None:
-        """When stdout has a result record with session_id, it wins over result.session_id."""
-        stdout = json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "result": "done",
-                "session_id": "stdout-uuid",
-                "is_error": False,
-            }
-        )
-        result = SubprocessResult(
-            returncode=0,
-            stdout=stdout,
-            stderr="",
-            termination=TerminationReason.NATURAL_EXIT,
-            pid=1,
-            session_id="ch-b-uuid",
-        )
-        sr = _build_skill_result(result)
-        assert sr.session_id == "stdout-uuid"
-
-    def test_context_exhaustion_session_id_from_subprocess_result(self) -> None:
-        """Context-exhaustion (no result record): uses result.session_id as fallback."""
-        partial_assistant_stdout = json.dumps(
-            {
-                "type": "assistant",
-                "message": {"content": [{"type": "text", "text": "Doing work..."}]},
-            }
-        )
-        result = SubprocessResult(
-            returncode=1,
-            stdout=partial_assistant_stdout,
-            stderr="",
-            termination=TerminationReason.NATURAL_EXIT,
-            pid=1,
-            session_id="ch-b-uuid-5678",
-        )
-        sr = _build_skill_result(result)
-        assert sr.session_id == "ch-b-uuid-5678"
-
-
-class TestHeadlessExecutorCompletionMarker:
-    """Protocol conformance: DefaultHeadlessExecutor.run accepts completion_marker."""
-
-    def test_headless_executor_accepts_completion_marker(self) -> None:
-        import inspect
-
-        from autoskillit.execution.headless import DefaultHeadlessExecutor
-
-        sig = inspect.signature(DefaultHeadlessExecutor.run)
-        assert "completion_marker" in sig.parameters
-        param = sig.parameters["completion_marker"]
-        assert param.default == ""
-
-
-class TestHeadlessExecutorIdleOutputTimeout:
-    """Protocol conformance and resolution logic for idle_output_timeout."""
-
-    def test_headless_executor_accepts_idle_output_timeout(self) -> None:
-        import inspect
-
-        from autoskillit.execution.headless import DefaultHeadlessExecutor
-
-        sig = inspect.signature(DefaultHeadlessExecutor.run)
-        assert "idle_output_timeout" in sig.parameters
-        param = sig.parameters["idle_output_timeout"]
-        assert param.default is None
-
-    def _success_payload(self, marker: str) -> SubprocessResult:
-        payload = json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "is_error": False,
-                "result": f"Done. {marker}",
-                "session_id": "sess-iot",
-            }
-        )
-        return SubprocessResult(0, payload, "", TerminationReason.NATURAL_EXIT, pid=1)
-
-    @pytest.mark.anyio
-    async def test_default_headless_executor_uses_per_step_idle_output_timeout(
-        self, tool_ctx
-    ) -> None:
-        """idle_output_timeout=120 is converted to float and passed to the runner."""
-        from autoskillit.execution.headless import run_headless_core
-
-        marker = tool_ctx.config.run_skill.completion_marker
-        tool_ctx.runner.push(self._success_payload(marker))
-        await run_headless_core(
-            "/investigate foo", cwd="/tmp", ctx=tool_ctx, idle_output_timeout=120.0
-        )
-        _, _cwd, _timeout, kwargs = tool_ctx.runner.call_args_list[0]
-        assert kwargs["idle_output_timeout"] == 120.0
-
-    @pytest.mark.anyio
-    async def test_default_headless_executor_converts_zero_to_none(self, tool_ctx) -> None:
-        """idle_output_timeout=0 is converted to None (disabled) before passing to runner."""
-        from autoskillit.execution.headless import run_headless_core
-
-        marker = tool_ctx.config.run_skill.completion_marker
-        tool_ctx.runner.push(self._success_payload(marker))
-        await run_headless_core(
-            "/investigate foo", cwd="/tmp", ctx=tool_ctx, idle_output_timeout=0.0
-        )
-        _, _cwd, _timeout, kwargs = tool_ctx.runner.call_args_list[0]
-        assert kwargs["idle_output_timeout"] is None
-
-    @pytest.mark.anyio
-    async def test_default_headless_executor_falls_back_to_cfg_idle_output_timeout(
-        self, tool_ctx
-    ) -> None:
-        """idle_output_timeout=None falls back to float(cfg.idle_output_timeout)."""
-        from autoskillit.execution.headless import run_headless_core
-
-        marker = tool_ctx.config.run_skill.completion_marker
-        tool_ctx.runner.push(self._success_payload(marker))
-        await run_headless_core(
-            "/investigate foo", cwd="/tmp", ctx=tool_ctx, idle_output_timeout=None
-        )
-        _, _cwd, _timeout, kwargs = tool_ctx.runner.call_args_list[0]
-        assert kwargs["idle_output_timeout"] == 600.0
-
-
-def _ndjson_with_write(result_text: str, file_paths: list[str], session_id: str = "test-session"):
-    """Build NDJSON stdout with Write tool_use entries and a result record."""
-    records = []
-    if file_paths:
-        content = [
-            {
-                "type": "tool_use",
-                "name": "Write",
-                "id": f"t{i}",
-                "input": {"file_path": fp},
-            }
-            for i, fp in enumerate(file_paths)
-        ]
-        records.append(json.dumps({"type": "assistant", "message": {"content": content}}))
-    records.append(
-        json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "is_error": False,
-                "result": result_text,
-                "session_id": session_id,
-            }
-        )
-    )
-    return "\n".join(records)
-
-
-class TestExtractMissingTokenHints:
-    def test_extracts_token_and_path(self):
-        stdout = _ndjson_with_write("plan summary\n%%ORDER_UP%%", ["/tmp/out.md"])
-        hints = _extract_missing_token_hints(stdout, [r"plan_path\s*=\s*/.+"])
-        assert hints == [("plan_path", "/tmp/out.md")]
-
-    def test_returns_empty_when_pattern_satisfied(self):
-        stdout = _ndjson_with_write("plan_path = /tmp/out.md\n%%ORDER_UP%%", ["/tmp/out.md"])
-        hints = _extract_missing_token_hints(stdout, [r"plan_path\s*=\s*/.+"])
-        assert hints == []
-
-    def test_returns_empty_for_non_path_patterns(self):
-        stdout = _ndjson_with_write("%%ORDER_UP%%", ["/tmp/out.md"])
-        hints = _extract_missing_token_hints(stdout, [r"verdict\s*=\s*\w+"])
-        assert hints == []
-
-    def test_uses_last_write_path(self):
-        stdout = _ndjson_with_write("%%ORDER_UP%%", ["/tmp/first.md", "/tmp/final.md"])
-        hints = _extract_missing_token_hints(stdout, [r"plan_path\s*=\s*/.+"])
-        assert hints == [("plan_path", "/tmp/final.md")]
-
-
-class TestContractNudge:
-    """Integration tests for the contract recovery nudge in run_headless_core.
-
-    The nudge fires when CONTRACT_RECOVERY is triggered with a valid session_id.
-    CONTRACT_RECOVERY requires synthesis to have failed first. For CHANNEL_A/B
-    sessions, synthesis is skipped (gated on UNMONITORED), so Write evidence
-    with file_path triggers CONTRACT_RECOVERY while _extract_missing_token_hints
-    can still find the file_path for hints.
-    """
-
-    def _main_session_ndjson(
-        self, marker: str, *, include_token: bool = False, session_id: str = "sess-main"
-    ) -> str:
-        """Build NDJSON for a CHANNEL_A-confirmed session with Write evidence."""
-        result_text = "plan summary\n"
-        if include_token:
-            result_text += "plan_path = /tmp/out.md\n"
-        result_text += marker
-        return _ndjson_with_write(result_text, ["/tmp/out.md"], session_id=session_id)
-
-    def _main_subprocess_result(
-        self, marker: str, *, session_id: str = "sess-main"
-    ) -> SubprocessResult:
-        """Build a SubprocessResult with CHANNEL_A confirmation to bypass synthesis."""
-        return SubprocessResult(
-            returncode=0,
-            stdout=self._main_session_ndjson(marker, session_id=session_id),
-            stderr="",
-            termination=TerminationReason.NATURAL_EXIT,
-            pid=1,
-            channel_confirmation=ChannelConfirmation.CHANNEL_A,
-        )
-
-    def _nudge_response_ndjson(self, marker: str, *, include_token: bool = True) -> str:
-        """Build NDJSON for a nudge response."""
-        if include_token:
-            result_text = f"plan_path = /tmp/out.md\n{marker}"
-        else:
-            result_text = f"I cannot do that.\n{marker}"
-        return json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "is_error": False,
-                "result": result_text,
-                "session_id": "sess-main",
-                "usage": {"input_tokens": 100, "output_tokens": 50},
-            }
-        )
-
-    @pytest.mark.anyio
-    async def test_nudge_fires_on_contract_recovery(self, tool_ctx):
-        from autoskillit.execution.headless import run_headless_core
-
-        marker = tool_ctx.config.run_skill.completion_marker
-        tool_ctx.runner.push(self._main_subprocess_result(marker))
-        tool_ctx.runner.push(
-            SubprocessResult(
-                0, self._nudge_response_ndjson(marker), "", TerminationReason.NATURAL_EXIT, pid=2
-            )
-        )
-        result = await run_headless_core(
-            "/autoskillit:make-plan foo",
-            cwd="/tmp",
-            ctx=tool_ctx,
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-        )
-        assert result.success is True
-        assert result.needs_retry is False
-        assert "plan_path = /tmp/out.md" in result.result
-
-    @pytest.mark.anyio
-    async def test_nudge_failure_falls_through(self, tool_ctx):
-        from autoskillit.execution.headless import run_headless_core
-
-        marker = tool_ctx.config.run_skill.completion_marker
-        tool_ctx.runner.push(self._main_subprocess_result(marker))
-        tool_ctx.runner.push(
-            SubprocessResult(
-                0,
-                self._nudge_response_ndjson(marker, include_token=False),
-                "",
-                TerminationReason.NATURAL_EXIT,
-                pid=2,
-            )
-        )
-        result = await run_headless_core(
-            "/autoskillit:make-plan foo",
-            cwd="/tmp",
-            ctx=tool_ctx,
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-        )
-        assert result.retry_reason == RetryReason.CONTRACT_RECOVERY
-        assert result.needs_retry is True
-
-    @pytest.mark.anyio
-    async def test_nudge_timeout_falls_through(self, tool_ctx):
-        from autoskillit.execution.headless import run_headless_core
-
-        marker = tool_ctx.config.run_skill.completion_marker
-        tool_ctx.runner.push(self._main_subprocess_result(marker))
-        tool_ctx.runner.push(
-            SubprocessResult(1, "", "timeout", TerminationReason.TIMED_OUT, pid=2)
-        )
-        result = await run_headless_core(
-            "/autoskillit:make-plan foo",
-            cwd="/tmp",
-            ctx=tool_ctx,
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-        )
-        assert result.retry_reason == RetryReason.CONTRACT_RECOVERY
-        assert result.needs_retry is True
-
-    @pytest.mark.anyio
-    async def test_nudge_skips_when_budget_exhausted(self, tool_ctx):
-        from datetime import UTC, datetime
-
-        from autoskillit.core import FailureRecord
-        from autoskillit.execution.headless import run_headless_core
-        from autoskillit.pipeline.audit import DefaultAuditLog
-
-        marker = tool_ctx.config.run_skill.completion_marker
-        audit = DefaultAuditLog()
-        for _ in range(4):
-            audit.record_failure(
-                FailureRecord(  # type: ignore[arg-type]
-                    timestamp=datetime.now(UTC).isoformat(),
-                    skill_command="/autoskillit:make-plan foo",
-                    exit_code=-1,
-                    subtype="stale",
-                    needs_retry=True,
-                    retry_reason="stale",
-                    stderr="",
-                )
-            )
-        tool_ctx.audit = audit
-        tool_ctx.runner.push(self._main_subprocess_result(marker))
-        result = await run_headless_core(
-            "/autoskillit:make-plan foo",
-            cwd="/tmp",
-            ctx=tool_ctx,
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-        )
-        assert result.retry_reason == RetryReason.BUDGET_EXHAUSTED
-        assert len(tool_ctx.runner.call_args_list) == 1
-
-    @pytest.mark.anyio
-    async def test_nudge_skips_when_no_session_id(self, tool_ctx):
-        from autoskillit.execution.headless import run_headless_core
-
-        marker = tool_ctx.config.run_skill.completion_marker
-        ndjson = _ndjson_with_write(f"plan summary\n{marker}", ["/tmp/out.md"], session_id="")
-        tool_ctx.runner.push(
-            SubprocessResult(
-                0,
-                ndjson,
-                "",
-                TerminationReason.NATURAL_EXIT,
-                pid=1,
-                channel_confirmation=ChannelConfirmation.CHANNEL_A,
-            )
-        )
-        result = await run_headless_core(
-            "/autoskillit:make-plan foo",
-            cwd="/tmp",
-            ctx=tool_ctx,
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-        )
-        assert result.retry_reason == RetryReason.CONTRACT_RECOVERY
-        assert len(tool_ctx.runner.call_args_list) == 1
-
-    @pytest.mark.anyio
-    async def test_nudge_resume_cmd_uses_correct_session_id(self, tool_ctx):
-        from autoskillit.execution.headless import run_headless_core
-
-        marker = tool_ctx.config.run_skill.completion_marker
-        tool_ctx.runner.push(self._main_subprocess_result(marker, session_id="sess-abc-123"))
-        tool_ctx.runner.push(
-            SubprocessResult(
-                0, self._nudge_response_ndjson(marker), "", TerminationReason.NATURAL_EXIT, pid=2
-            )
-        )
-        await run_headless_core(
-            "/autoskillit:make-plan foo",
-            cwd="/tmp",
-            ctx=tool_ctx,
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-        )
-        assert len(tool_ctx.runner.call_args_list) == 2
-        nudge_cmd = tool_ctx.runner.call_args_list[1][0]
-        assert "--resume" in nudge_cmd
-        resume_idx = nudge_cmd.index("--resume")
-        assert nudge_cmd[resume_idx + 1] == "sess-abc-123"
-
-    @pytest.mark.anyio
-    async def test_nudge_token_usage_merged(self, tool_ctx):
-        from autoskillit.execution.headless import run_headless_core
-
-        marker = tool_ctx.config.run_skill.completion_marker
-        main_ndjson_records = [
-            json.dumps(
-                {
-                    "type": "assistant",
-                    "message": {
-                        "content": [
-                            {
-                                "type": "tool_use",
-                                "name": "Write",
-                                "id": "t0",
-                                "input": {"file_path": "/tmp/out.md"},
-                            }
-                        ]
-                    },
-                }
-            ),
-            json.dumps(
-                {
-                    "type": "result",
-                    "subtype": "success",
-                    "is_error": False,
-                    "result": f"plan summary\n{marker}",
-                    "session_id": "sess-tok",
-                    "usage": {"input_tokens": 1000, "output_tokens": 500},
-                }
-            ),
-        ]
-        main_ndjson = "\n".join(main_ndjson_records)
-
-        nudge_ndjson = json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "is_error": False,
-                "result": f"plan_path = /tmp/out.md\n{marker}",
-                "session_id": "sess-tok",
-                "usage": {"input_tokens": 200, "output_tokens": 100},
-            }
-        )
-
-        tool_ctx.runner.push(
-            SubprocessResult(
-                0,
-                main_ndjson,
-                "",
-                TerminationReason.NATURAL_EXIT,
-                pid=1,
-                channel_confirmation=ChannelConfirmation.CHANNEL_A,
-            )
-        )
-        tool_ctx.runner.push(
-            SubprocessResult(0, nudge_ndjson, "", TerminationReason.NATURAL_EXIT, pid=2)
-        )
-
-        result = await run_headless_core(
-            "/autoskillit:make-plan foo",
-            cwd="/tmp",
-            ctx=tool_ctx,
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-            step_name="test-step",
-        )
-        assert result.success is True
-        assert result.token_usage is not None
-        assert result.token_usage.get("input_tokens", 0) >= 1200
-        assert result.token_usage.get("output_tokens", 0) >= 600
-
-    @pytest.mark.anyio
-    async def test_nudge_exception_falls_through(self, tool_ctx):
-        from autoskillit.execution.headless import run_headless_core
-
-        marker = tool_ctx.config.run_skill.completion_marker
-        tool_ctx.runner.push(self._main_subprocess_result(marker))
-        # Empty stdout nudge → patterns not found → falls through
-        tool_ctx.runner.push(
-            SubprocessResult(1, "", "RuntimeError", TerminationReason.NATURAL_EXIT, pid=2)
-        )
-        result = await run_headless_core(
-            "/autoskillit:make-plan foo",
-            cwd="/tmp",
-            ctx=tool_ctx,
-            expected_output_patterns=[r"plan_path\s*=\s*/.+"],
-        )
-        assert result.retry_reason == RetryReason.CONTRACT_RECOVERY
-        assert result.needs_retry is True
-
-
-# ---------------------------------------------------------------------------
-# last_stop_reason threading test
-# ---------------------------------------------------------------------------
-
-
-def test_build_skill_result_surfaces_last_stop_reason():
-    ndjson = "\n".join(
-        [
-            json.dumps(
-                {
-                    "type": "assistant",
-                    "message": {"stop_reason": "end_turn", "content": [], "usage": {}},
-                }
-            ),
-            json.dumps(
-                {
-                    "type": "result",
-                    "subtype": "success",
-                    "is_error": False,
-                    "result": "done",
-                    "session_id": "s1",
-                }
-            ),
-        ]
-    )
-    result = _make_result(returncode=0, stdout=ndjson)
-    sr = _build_skill_result(result)
-    assert sr.last_stop_reason == "end_turn"
-
-
-class TestDispatchFoodTruck:
-    """Tests for DefaultHeadlessExecutor.dispatch_food_truck."""
-
-    def _make_success_stdout(self, marker: str = "%%FT_DONE%%") -> str:
-        import json
-
-        return json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "result": f"L2 done {marker}",
-                "session_id": "ft-session",
-                "is_error": False,
-            }
-        )
-
-    @pytest.mark.anyio
-    async def test_dispatch_food_truck_calls_runner(self, minimal_ctx, tmp_path: Path):
-        from autoskillit.core.types import SubprocessResult, TerminationReason
-        from autoskillit.execution.headless import DefaultHeadlessExecutor
-        from tests.fakes import MockSubprocessRunner
-
-        runner = MockSubprocessRunner()
-        runner.set_default(
-            SubprocessResult(
-                returncode=0,
-                stdout=self._make_success_stdout(),
-                stderr="",
-                termination=TerminationReason.NATURAL_EXIT,
-                pid=55555,
-            )
-        )
-        minimal_ctx.runner = runner
-        minimal_ctx.plugin_source = DirectInstall(plugin_dir=tmp_path)
-
-        executor = DefaultHeadlessExecutor(minimal_ctx)
-        await executor.dispatch_food_truck(
-            "You are an L2 orchestrator",
-            str(tmp_path),
-            completion_marker="%%FT_DONE%%",
-        )
-
-        assert runner.call_args_list, "runner was never called"
-        cmd, _cwd, _timeout, kwargs = runner.call_args_list[0]
-        env = kwargs.get("env")
-        assert env is not None
-        assert env["AUTOSKILLIT_SESSION_TYPE"] == "orchestrator"
-        assert env["AUTOSKILLIT_HEADLESS"] == "1"
-        assert "--tools" in cmd
-        assert "AskUserQuestion" in cmd
-
-    @pytest.mark.anyio
-    async def test_dispatch_food_truck_returns_skill_result(self, minimal_ctx, tmp_path: Path):
-        from autoskillit.core.types import SkillResult, SubprocessResult, TerminationReason
-        from autoskillit.execution.headless import DefaultHeadlessExecutor
-        from tests.fakes import MockSubprocessRunner
-
-        runner = MockSubprocessRunner()
-        runner.set_default(
-            SubprocessResult(
-                returncode=0,
-                stdout=self._make_success_stdout(),
-                stderr="",
-                termination=TerminationReason.NATURAL_EXIT,
-                pid=55555,
-            )
-        )
-        minimal_ctx.runner = runner
-        minimal_ctx.plugin_source = DirectInstall(plugin_dir=tmp_path)
-
-        executor = DefaultHeadlessExecutor(minimal_ctx)
-        result = await executor.dispatch_food_truck(
-            "You are an L2 orchestrator",
-            str(tmp_path),
-            completion_marker="%%FT_DONE%%",
-        )
-
-        assert isinstance(result, SkillResult)
-        assert result.success is True
-
-    @pytest.mark.anyio
-    async def test_dispatch_food_truck_on_spawn_receives_pid(self, minimal_ctx, tmp_path: Path):
-        from autoskillit.core.types import SubprocessResult, TerminationReason
-        from autoskillit.execution.headless import DefaultHeadlessExecutor
-        from tests.fakes import MockSubprocessRunner
-
-        runner = MockSubprocessRunner()
-        runner.set_default(
-            SubprocessResult(
-                returncode=0,
-                stdout=self._make_success_stdout(),
-                stderr="",
-                termination=TerminationReason.NATURAL_EXIT,
-                pid=55555,
-            )
-        )
-        minimal_ctx.runner = runner
-        minimal_ctx.plugin_source = DirectInstall(plugin_dir=tmp_path)
-
-        spawned_pids: list[int] = []
-
-        executor = DefaultHeadlessExecutor(minimal_ctx)
-        await executor.dispatch_food_truck(
-            "You are an L2 orchestrator",
-            str(tmp_path),
-            completion_marker="%%FT_DONE%%",
-            on_spawn=lambda pid, ticks: spawned_pids.append(pid),
-        )
-
-        assert spawned_pids == [55555]
-
-    @pytest.mark.anyio
-    async def test_dispatch_food_truck_on_spawn_not_required(self, minimal_ctx, tmp_path: Path):
-        from autoskillit.core.types import SkillResult, SubprocessResult, TerminationReason
-        from autoskillit.execution.headless import DefaultHeadlessExecutor
-        from tests.fakes import MockSubprocessRunner
-
-        runner = MockSubprocessRunner()
-        runner.set_default(
-            SubprocessResult(
-                returncode=0,
-                stdout=self._make_success_stdout(),
-                stderr="",
-                termination=TerminationReason.NATURAL_EXIT,
-                pid=55555,
-            )
-        )
-        minimal_ctx.runner = runner
-        minimal_ctx.plugin_source = DirectInstall(plugin_dir=tmp_path)
-
-        executor = DefaultHeadlessExecutor(minimal_ctx)
-        result = await executor.dispatch_food_truck(
-            "You are an L2 orchestrator",
-            str(tmp_path),
-            completion_marker="%%FT_DONE%%",
-            on_spawn=None,
-        )
-
-        assert isinstance(result, SkillResult)
-        assert result.success is True
-
-    @pytest.mark.anyio
-    async def test_dispatch_food_truck_marketplace_install_does_not_raise(
-        self, minimal_ctx, tmp_path: Path
-    ):
-        """dispatch_food_truck with MarketplaceInstall resolves cache_path — no ValueError."""
-        from autoskillit.core.types import SubprocessResult, TerminationReason
-        from autoskillit.execution.headless import DefaultHeadlessExecutor
-        from tests.fakes import MockSubprocessRunner
-
-        cache = tmp_path / "marketplace_cache"
-        cache.mkdir()
-        runner = MockSubprocessRunner()
-        runner.set_default(
-            SubprocessResult(
-                returncode=0,
-                stdout=self._make_success_stdout(),
-                stderr="",
-                termination=TerminationReason.NATURAL_EXIT,
-                pid=55555,
-            )
-        )
-        minimal_ctx.runner = runner
-        minimal_ctx.plugin_source = MarketplaceInstall(cache_path=cache)
-
-        executor = DefaultHeadlessExecutor(minimal_ctx)
-        await executor.dispatch_food_truck(
-            "You are an L2 orchestrator",
-            str(tmp_path),
-            completion_marker="%%FT_DONE%%",
-        )
-        cmd, _cwd, _timeout, _kwargs = runner.call_args_list[0]
-        assert "--plugin-dir" in cmd
-        assert str(cache) in cmd
-
-
-class TestDispatchFoodTruckPackInjection:
-    """Tests that dispatch_food_truck correctly injects AUTOSKILLIT_L2_TOOL_TAGS."""
-
-    def _make_success_stdout(self, marker: str = "%%FT_DONE%%") -> str:
-        import json
-
-        return json.dumps(
-            {
-                "type": "result",
-                "subtype": "success",
-                "result": f"L2 done {marker}",
-                "session_id": "ft-session",
-                "is_error": False,
-            }
-        )
-
-    @pytest.mark.anyio
-    async def test_requires_packs_injected_as_l2_tool_tags(self, minimal_ctx, tmp_path: Path):
-        """dispatch_food_truck with requires_packs injects sorted comma-joined env var."""
-        from autoskillit.core.types import SubprocessResult, TerminationReason
-        from autoskillit.execution.headless import DefaultHeadlessExecutor
-        from tests.fakes import MockSubprocessRunner
-
-        runner = MockSubprocessRunner()
-        runner.set_default(
-            SubprocessResult(
-                returncode=0,
-                stdout=self._make_success_stdout(),
-                stderr="",
-                termination=TerminationReason.NATURAL_EXIT,
-                pid=12345,
-            )
-        )
-        minimal_ctx.runner = runner
-        minimal_ctx.plugin_source = DirectInstall(plugin_dir=tmp_path)
-
-        executor = DefaultHeadlessExecutor(minimal_ctx)
-        await executor.dispatch_food_truck(
-            "You are an L2 orchestrator",
-            str(tmp_path),
-            completion_marker="%%FT_DONE%%",
-            requires_packs=["ci", "github", "clone"],
-        )
-
-        assert runner.call_args_list, "runner was never called"
-        _cmd, _cwd, _timeout, kwargs = runner.call_args_list[0]
-        env = kwargs.get("env")
-        assert env is not None
-        assert env["AUTOSKILLIT_L2_TOOL_TAGS"] == "ci,clone,github"
-
-    @pytest.mark.anyio
-    async def test_requires_packs_empty_omits_l2_tool_tags(self, minimal_ctx, tmp_path: Path):
-        """dispatch_food_truck with empty requires_packs does not inject L2_TOOL_TAGS."""
-        from autoskillit.core.types import SubprocessResult, TerminationReason
-        from autoskillit.execution.headless import DefaultHeadlessExecutor
-        from tests.fakes import MockSubprocessRunner
-
-        runner = MockSubprocessRunner()
-        runner.set_default(
-            SubprocessResult(
-                returncode=0,
-                stdout=self._make_success_stdout(),
-                stderr="",
-                termination=TerminationReason.NATURAL_EXIT,
-                pid=12345,
-            )
-        )
-        minimal_ctx.runner = runner
-        minimal_ctx.plugin_source = DirectInstall(plugin_dir=tmp_path)
-
-        executor = DefaultHeadlessExecutor(minimal_ctx)
-        await executor.dispatch_food_truck(
-            "You are an L2 orchestrator",
-            str(tmp_path),
-            completion_marker="%%FT_DONE%%",
-            requires_packs=[],
-        )
-
-        assert runner.call_args_list, "runner was never called"
-        _cmd, _cwd, _timeout, kwargs = runner.call_args_list[0]
-        env = kwargs.get("env")
-        assert env is not None
-        assert "AUTOSKILLIT_L2_TOOL_TAGS" not in env
-
-
-def test_default_executor_satisfies_protocol_with_dispatch(minimal_ctx) -> None:
-    """DefaultHeadlessExecutor satisfies HeadlessExecutor protocol with dispatch_food_truck."""
-    from autoskillit.core import HeadlessExecutor
-    from autoskillit.execution.headless import DefaultHeadlessExecutor
-
-    executor = DefaultHeadlessExecutor(minimal_ctx)
-    assert isinstance(executor, HeadlessExecutor)
